% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5File.R
\docType{data}
\name{H5File}
\alias{H5File}
\title{A Place Holder for the hdf5r::H5File R6 Class}
\format{An object of class \code{list} of length 1.}
\usage{
H5File
}
\description{
A Place Holder for the hdf5r::H5File R6 Class
}
\keyword{datasets}
