% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDFile.R
\docType{data}
\name{HDFile}
\alias{HDFile}
\title{A Place Holder for the hdf5r::HDFile R6 Class}
\format{An object of class \code{list} of length 1.}
\usage{
HDFile
}
\description{
A Place Holder for the hdf5r::HDFile R6 Class
}
\keyword{datasets}
